LATTICE-2417 move and wait
