+++ src/main/kotlin/com/openlattice/users/Auth0UserListingService.kt
@@ -23,9 +23,18 @@ package com.openlattice.users
 
 import com.auth0.client.mgmt.ManagementAPI
 import com.auth0.json.mgmt.users.User
+import com.dataloom.mappers.ObjectMappers
+import com.geekbeast.auth0.*
 import com.openlattice.users.export.Auth0ApiExtension
+import com.openlattice.users.export.JobStatus
+import com.openlattice.users.export.UserExportJobRequest
+import com.openlattice.users.export.UserExportJobResult
 import org.slf4j.LoggerFactory
+import java.io.BufferedReader
+import java.io.InputStreamReader
 import java.time.Instant
+import java.util.stream.Collectors
+import java.util.zip.GZIPInputStream
 
 const val SEARCH_ENGINE_VERSION = "v3"
 const val MAX_PAGE_SIZE = 100
@@ -39,11 +48,60 @@ class Auth0UserListingService(
         private val managementApi: ManagementAPI, private val auth0ApiExtension: Auth0ApiExtension
 ) : UserListingService {
 
+    companion object {
+        private val logger = LoggerFactory.getLogger(Auth0UserListingService::class.java)
+    }
+
     /**
      * Retrieves all users from auth0 as a result of an export job.
      */
     override fun getAllUsers(): Sequence<User> {
         return getUsers(auth0ApiExtension).asSequence()
+    }
+
+    /**
+     * Calls an export job to download all users from auth0 to a json and parses it to a sequence of users.
+     *
+     * Note: This export is used, because the user API has a 1000 user limitation.
+     * @see <a href="https://auth0.com/docs/users/search/v3/get-users-endpoint#limitations"> Auth0 user endpoint
+     * limitations </a>
+     */
+    private fun getUsers(auth0ApiExtension: Auth0ApiExtension): List<User> {
+        val exportEntity = auth0ApiExtension.userExport()
+        val job = exportEntity.submitExportJob(
+                UserExportJobRequest(listOf(USER_ID, EMAIL, NICKNAME, APP_METADATA, IDENTITIES))
+        )
+
+        // will fail if export job hangs too long with too many requests error (429 status code)
+        // https://auth0.com/docs/policies/rate-limits
+        var exportJobResult = exportEntity.getJob(job.id)
+        while (exportJobResult.status == JobStatus.PENDING) {
+            Thread.sleep(200) // wait before calling again for job status
+            exportJobResult = exportEntity.getJob(job.id)
+        }
+
+        return readUsersFromLocation(exportJobResult)
+    }
+
+    private fun readUsersFromLocation(exportJobResult: UserExportJobResult): List<User> {
+        val downloadUrl = exportJobResult.location.get()
+
+        try {
+            val mapper = ObjectMappers.getMapper(ObjectMappers.Mapper.valueOf(exportJobResult.format.name.toUpperCase()))
+
+            val connection = downloadUrl.openConnection()
+            connection.setRequestProperty("Accept-Encoding", "gzip")
+
+            val input = GZIPInputStream(connection.getInputStream())
+            val buffered = BufferedReader(InputStreamReader(input, "UTF-8"))
+
+            // export json format has a line by line user object format
+            // if at any point we have too many users, we might have to download the file
+            return buffered.lines().map { line -> mapper.readValue(line, User::class.java) }.collect(Collectors.toList())
+        } catch (e: Exception) {
+            logger.error("Couldn't read list of users from download url $downloadUrl.")
+            throw e
+        }
     }
 
     /**
